+++ R/helpers.R
@@ -46,18 +46,6 @@ is.historical <- function(.data) { # nolint: object_name_linter
 }
 
 
-
-#' not-in operator
-#' @inheritParams base::`%in%`
-#' @examples
-#' 2 %notin% c(1,3) # TRUE
-#' @return A logical vector, indicating if a match was NOT located for each element of x
-#' @export
-`%notin%` <- function(x, table) {
-  return(!(x %in% table))
-}
-
-
 #' checkmate helper: Assert "generic" data.table/data.frame/tbl/tibble type
 #' @param .data Object to test if is data.table, data.frame, tbl or tibble
 #' @param ...   Parameters passed to checkmate::check_*